+++ ddl/postgres/i2b2demodata/modifier_dimension.sql
@@ -10,8 +10,8 @@ CREATE TABLE modifier_dimension (
     download_date timestamp without time zone,
     import_date timestamp without time zone,
     sourcesystem_cd character varying(50),
-    upload_id numeric(38,0),
-    modifier_level numeric(18,0),
+    upload_id bigint,
+    modifier_level bigint,
     modifier_node_type character varying(10)
 );
